Tweaking the sidebar
Making the rows a little larger to keep them in line with the default
theme.  Also decreasing the indent offset so the space between the
close button, wall, and label seems a little more equal.

Also setting dark_content to true, though I'm not sure if that has any
effect whatsoever.