Fix bug with tabs and window-shrinking.
If the window was resized in a certain way, the previous tabs could have
caused some sort of odd crash in ST2. So, the tabs are slightly modified
to prevent this. They're also probably visually a little different.

Also, fixed an odd 1px column missing from the active modified tabs in
Nil.